+++ R/geom-debug.R
@@ -62,8 +62,11 @@
 #'     geom_null()
 #' }
 #'
-geom_null <- function(mapping = NULL, data = NULL, stat = "identity",
-                      position = "identity", na.rm = FALSE,
+geom_null <- function(mapping = NULL,
+                      data = NULL,
+                      stat = "identity",
+                      position = "identity",
+                      na.rm = FALSE,
                       show.legend = FALSE,
                       inherit.aes = TRUE, ...) {
   ggplot2::layer(
@@ -88,35 +91,47 @@ geom_null <- function(mapping = NULL, data = NULL, stat = "identity",
 #'
 GeomNull <-
   ggplot2::ggproto("GeomNull", ggplot2::Geom,
-                   optional_aes = c(# ggplot2
-                     "x", "y", "npcx", "npcy",
+                   # needed to avoid warnings
+                   optional_aes = c(# ggplot2 at 3.5.1
+                     "group", "subgroup", "order",
+                     "x", "y", "z",
                      "xintercept", "yintercept",
+                     "slope", "intercept",
+                     "sample",
                      "xend", "yend",
                      "ymax", "ymin",
                      "xmax", "xmin",
                      "middle", "xmiddle",
                      "lower", "xlower",
                      "upper", "xupper",
-                     "width",
+                     "width", "height",
                      "label",
                      "hjust", "vjust",
                      "colour", "color", "alpha", "fill",
                      "shape", "size", "linewidth", "stroke",
                      "linetype", "lineend", "linejoin",
-                     "angle",
+                     "angle", "radius",
                      "family", "fontface", "lineheight",
                      "map_id", "weight",
-                     # ggpp and ggrepel
-                     "segment.linetype", "segment.colour",
-                     "segment.size", "segment.alpha",
-                     # ggpp
+                     "geometry",
+                     # ggpp at 0.5.6
                      "vp.width", "vp.height",
-                     # ggrepel
+                     "npcx", "npcy",
+                     # ggrepel at 0.9.5
                      "point.size", "segment.angle",
                      "segment.curvature", "segment.ncp",
                      "segment.shape", "segment.square",
                      "segment.squareShape", "segment.inflect",
-                     "segment.debug"
+                     "segment.debug",
+                     "segment.linetype",
+                     "segment.colour", "segment.color",
+                     "segment.size", "segment.alpha",
+                     # ggtext at 0.1.2
+                     "label.colour", "label.size",
+                     "text.colour",
+                     "box.colour", "box.size",
+                     "halign", "valign",
+                     "orientation"
                    ),
                    default_aes = ggplot2::aes(),
                    draw_key = function(...) {
@@ -129,15 +144,61 @@ GeomNull <-
                    }
   )
 
+# used only in the file
+# defined to avoid duplicating this list
+optional.aes <-
+  c(# ggplot2 at 3.5.1
+    "group", "subgroup", "order",
+    "x", "y", "z",
+    "xintercept", "yintercept",
+    "slope", "intercept",
+    "sample",
+    "xend", "yend",
+    "ymax", "ymin",
+    "xmax", "xmin",
+    "middle", "xmiddle",
+    "lower", "xlower",
+    "upper", "xupper",
+    "width", "height",
+    "label",
+    "hjust", "vjust",
+    "colour", "color", "alpha", "fill",
+    "shape", "size", "linewidth", "stroke",
+    "linetype", "lineend", "linejoin",
+    "angle", "radius",
+    "family", "fontface", "lineheight",
+    "map_id", "weight",
+    "geometry",
+    # ggpp at 0.5.6
+    "vp.width", "vp.height",
+    "npcx", "npcy",
+    # ggrepel at 0.9.5
+    "point.size", "segment.angle",
+    "segment.curvature", "segment.ncp",
+    "segment.shape", "segment.square",
+    "segment.squareShape", "segment.inflect",
+    "segment.debug",
+    "segment.linetype",
+    "segment.colour", "segment.color",
+    "segment.size", "segment.alpha",
+    # ggtext at 0.1.2
+    "label.colour", "label.size",
+    "text.colour",
+    "box.colour", "box.size",
+    "halign", "valign",
+    "orientation"
+  )
+
 # Debug geom --------------------------------------------------------------
 
-#' Geom that prints input data to console.
+#' Geoms that print input data to console.
 #'
-#' The debug geom is used to print to the console a summary of the data being
-#' received by geoms as input \code{data} data frame.
+#' The debug geoms are used to print to the console a summary of the object being
+#' received by geoms' draw functions as input \code{data} and \code{parameters}
+#' objects.
 #'
 #' It can be useful when debugging the code of statistics or to learn how the
-#' stats and geoms work in 'ggplot2' (>= 2.0.0).
+#' stats and geoms work in 'ggplot2' (>= 3.0.0).
 #'
 #' @param mapping Set of aesthetic mappings created by
 #'   \code{\link[ggplot2]{aes}} or \code{\link{aes_}}. If specified and
@@ -146,11 +207,15 @@ GeomNull <-
 #'   \code{mapping} if there isn't a mapping defined for the plot.
 #' @param data A data frame. If specified, overrides the default data frame
 #'   defined at the top level of the plot.
-#' @param summary.fun The name of a function as a character string (or a
-#'   function) to be used to summarize the \code{data} object received as input
-#'   before printing it.
-#' @param summary.fun.args A list of additional arguments
-#'   to be passed to \code{summary.fun}.
+#' @param dbgfun.data,dbgfun.params The functions as character strings giving
+#'   their names or as named or anonymous function objects, to be used to
+#'   summarize the \code{data} and the \code{params} objects received as input
+#'   by the geometry.
+#' @param dbgfun.data.args,dbgfun.params.args A named list of additional
+#'   arguments to be passed to \code{dbgfun.data} and \code{dbgfun.params}.
+#' @param dbgfun.print A function used to print the \code{data} object received
+#'   as input.
+#' @param dbgfun.print.args A named list.
 #' @param position Position adjustment, either as a string, or the result of a
 #'   call to a position adjustment function.
 #' @param stat The statistical transformation to use on the data for this layer,
@@ -164,7 +229,7 @@ GeomNull <-
 #'   than combining with them. This is most useful for helper functions that
 #'   define both data and aesthetics and shouldn't inherit behaviour from the
 #'   default plot specification, e.g. \code{\link[ggplot2]{borders}}.
-#' @param parse Ignored. Helps avoid warnings.
+#' @param parse,orientation Ignored. Helps avoid warnings.
 #' @param nudge_x,nudge_y Horizontal and vertical adjustments to nudge the
 #'   starting position. The units for \code{nudge_x} and \code{nudge_y} are the
 #'   same as for the data units on the x-axis and y-axis.
@@ -182,28 +247,45 @@ GeomNull <-
 #'   effect.
 #'
 #' @details The intended use of this geometry is to help explore the data as
-#'   they are added to a plot layer. It is very unusual in that it does not
-#'   produce visible graphic output. It only returns a \code{grid.null()} grob
-#'   (graphical object) when the plot is rendered. Also, differently to normal
-#'   geometries, it passes its \code{data} input as argument to the first
-#'   parameter of the function passed as argument to \code{summary.fun}. The
-#'   value returned by this function is then printed to the R console. If
-#'   \code{summary.fun = NULL}, the whole \code{data} object is printed.
+#'   they are used in a plot layer to produce graphical output. Geometries
+#'   can be defined using draw functions that receive as input data coresponding
+#'   to a single group at a time, or draw functions that receive as input all
+#'   data to be drawn in a panel at a time, possibly including multiple
+#'   groups.
+#'
+#'   These \emph{debug} geoms are very unusual in that they do not produce
+#'   visible graphic output. They "draw" a \code{grid.null()} grob (graphical
+#'   object) when the plot is rendered. Also, differently to normal geometries,
+#'   they print the \code{data} and \code{params} objects or a summary of them
+#'   to the R console. The summary is obtained using the functions passed as
+#'   arguments to their formal parameter \code{dbgfun.data} and
+#'   \code{dbgfun.params}. The \code{data} and \code{params} objects are passed
+#'   as the first positional argument to these functions and the values they
+#'   return are printed.
+#'
+#'   If \code{dbgfun.data = "I"} is passed, the \code{data} object is printed as
+#'   is. In contrast, if \code{dbgfun.data = NULL} is passed, the \code{data}
+#'   object summary and its printing are not skipped. The mechanism is identical
+#'   for \code{dbgfun.params} and \code{params}.
 #'
 #'   Nudging with \code{nudge_x} and \code{nudge_y} behave as in
 #'   \code{\link[ggplot2]{geom_text}}. Arguments passed to \code{position} are
 #'   obeyed. So the effects of positions are reflected in the \code{data} object
-#'   printed or summarized to the R console. The argument passed to \code{parse}
-#'   is currently ignored.
+#'   printed or summarized to the R console. The arguments passed to
+#'   \code{parse} and \code{orientation} are currently ignored.
 #'
 #'   Many aesthetics are defined as optional so that they are accepted silently
-#'   by \code{geom_debug()} and handled by 'ggplot2' as usual.
+#'   by \code{geom_debug()} and handled by 'ggplot2' as usual. Given the number
+#'   available extensions to 'ggplot2', it is likely that some are missing.
 #'
-#'   If the argument passed to \code{summary.fun()} is a character string, the
-#'   name will appear in the header of the printout. However, the function
-#'   must be available at the time the plot is rendered. If a function is
-#'   passed as argument, its definition will be saved as part of the \code{"gg"}
-#'   object.
+#'   The names of \code{dbgfun.data()} and \code{dbgfun.params()} are included
+#'   in the section headers of the printout, together with panels and groups.
+#'
+#'   In most cases, the definitions of the debug and print functions must be
+#'   available when the \code{"gg"} object is printed and the plot rendered.
+#'
+#' @note \code{geom_debug()} is a synonym of \code{geom_debug_panel()}, for
+#'   backwards compatibility. Not to be used in new code.
 #'
 #' @seealso To access data, scales and grobs in a built ggplot, see
 #'   \code{\link[ggplot2]{ggplot_build}}.
@@ -215,46 +297,73 @@ GeomNull <-
 #' # echo to the R console \code{data} as received by geoms
 #' ggplot(mtcars, aes(cyl, mpg, color = factor(cyl))) +
 #'   geom_point() +
-#'   geom_debug()
+#'   geom_debug_panel()
 #'
 #' ggplot(mtcars, aes(cyl, mpg, color = factor(cyl))) +
 #'   geom_point() +
-#'   geom_debug(summary.fun = "head", summary.fun.args = list(n = 3))
+#'   geom_debug_group()
 #'
 #' ggplot(mtcars, aes(cyl, mpg, color = factor(cyl))) +
 #'   geom_point() +
-#'   geom_debug(summary.fun = "nrow")
+#'   geom_debug_panel(dbgfun.params = NULL)
 #'
 #' ggplot(mtcars, aes(cyl, mpg, color = factor(cyl))) +
 #'   geom_point() +
-#'   geom_debug(summary.fun = "attributes")
+#'   geom_debug_panel(dbgfun.data = NULL)
+#'
+#' ggplot(mtcars, aes(cyl, mpg, color = factor(cyl))) +
+#'   geom_point() +
+#'   geom_debug_panel(dbgfun.data = "head", dbgfun.data.args = list(n = 3))
+#'
+#' ggplot(mtcars, aes(cyl, mpg, color = factor(cyl))) +
+#'   geom_point() +
+#'   geom_debug_panel(dbgfun.data = "nrow", dbgfun.params = "length")
+#'
+#' ggplot(mtcars, aes(cyl, mpg, color = factor(cyl))) +
+#'   geom_point() +
+#'   geom_debug_panel(dbgfun.data = "attributes", dbgfun.params = "attributes")
+#'
+#' ggplot(mtcars, aes(cyl, mpg, color = factor(cyl))) +
+#'   geom_point() +
+#'   geom_debug_panel(dbgfun.data = "I", dbgfun.params = NULL)
 #'
 #' # echo to the R console \code{data} as received by geoms
 #' ggplot(mtcars, aes(cyl, mpg, colour = factor(cyl))) +
 #'   stat_summary(fun.data = "mean_se") +
-#'   stat_summary(fun.data = "mean_se", geom = "debug", summary.fun = NULL)
+#'   stat_summary(geom = "debug_panel", fun.data = "mean_se")
+#'
+#' ggplot(mtcars, aes(cyl, mpg, colour = factor(cyl))) +
+#'   stat_summary(fun.data = "mean_se") +
+#'   stat_summary(geom = "debug_panel", fun.data = "mean_se", dbgfun.params = NULL)
 #'
 #' # shape data is not passed to geometries or statistics
 #' if (requireNamespace("sf", quietly = TRUE)) {
-#'   nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
+#'   nc <-
+#'     sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
 #'
 #'   ggplot(data = nc) +
 #'     geom_sf(color = "darkblue", fill = "white") +
-#'     geom_debug()
+#'     geom_debug_panel()
 #' }
 #'
-geom_debug <- function(mapping = NULL,
-                       data = NULL,
-                       stat = "identity",
-                       summary.fun = "head",
-                       summary.fun.args = list(),
-                       parse = NULL,
-                       nudge_x = 0,
-                       nudge_y = 0,
-                       position = "identity", na.rm = FALSE,
-                       show.legend = FALSE,
-                       inherit.aes = TRUE,
-                       ...) {
+geom_debug_panel <- function(mapping = NULL,
+                             data = NULL,
+                             stat = "identity",
+                             dbgfun.data = "head",
+                             dbgfun.data.args = list(),
+                             dbgfun.params = "summary",
+                             dbgfun.params.args = list(),
+                             dbgfun.print = "print",
+                             dbgfun.print.args = list(),
+                             parse = NULL,
+                             orientation = NULL,
+                             nudge_x = 0,
+                             nudge_y = 0,
+                             position = "identity",
+                             na.rm = FALSE,
+                             show.legend = FALSE,
+                             inherit.aes = TRUE,
+                             ...) {
   if (!missing(nudge_x) || !missing(nudge_y)) {
     if (!missing(position) && position != "identity") {
       rlang::abort("You must specify either `position` or `nudge_x`/`nudge_y`.")
@@ -264,7 +373,7 @@ geom_debug <- function(mapping = NULL,
   }
 
   ggplot2::layer(
-    geom = GeomDebug,
+    geom = GeomDebugPanel,
     mapping = mapping,
     data = data,
     stat = stat,
@@ -272,76 +381,156 @@ geom_debug <- function(mapping = NULL,
     show.legend = show.legend,
     inherit.aes = inherit.aes,
     params = rlang::list2(na.rm = na.rm,
-                          summary.fun = summary.fun,
-                          summary.fun.args = summary.fun.args,
-                          ...)
+                          ...,
+                          dbgfun.data = dbgfun.data,
+                          dbgfun.data.args = dbgfun.data.args,
+                          dbgfun.params = dbgfun.params,
+                          dbgfun.params.args = dbgfun.params.args,
+                          dbgfun.print = dbgfun.print,
+                          dbgfun.print.args = dbgfun.print.args,
+                          draw.label = "draw_panel()")
   )
 }
-
-#' @rdname geom_debug
-#' @export
-geom_debug_npc <- geom_debug
 
 #' @rdname gginnards-ggproto
 #' @format NULL
 #' @usage NULL
 #' @export
-GeomDebug <-
-  ggplot2::ggproto("GeomDebug", ggplot2::Geom,
-                   optional_aes = c(# ggplot2
-                                    "x", "y", "npcx", "npcy",
-                                    "xintercept", "yintercept",
-                                    "xend", "yend",
-                                    "ymax", "ymin",
-                                    "xmax", "xmin",
-                                    "middle", "xmiddle",
-                                    "lower", "xlower",
-                                    "upper", "xupper",
-                                    "width",
-                                    "label",
-                                    "hjust", "vjust",
-                                    "colour", "color", "alpha", "fill",
-                                    "shape", "size", "linewidth", "stroke",
-                                    "linetype", "lineend", "linejoin",
-                                    "angle",
-                                    "family", "fontface", "lineheight",
-                                    "map_id", "weight",
-                                    # ggpp and ggrepel
-                                    "segment.linetype", "segment.colour",
-                                    "segment.size", "segment.alpha",
-                                    # ggpp
-                                    "vp.width", "vp.height",
-                                    # ggrepel
-                                    "point.size", "segment.angle",
-                                    "segment.curvature", "segment.ncp",
-                                    "segment.shape", "segment.square",
-                                    "segment.squareShape", "segment.inflect",
-                                    "segment.debug"
-                                    ),
-                   default_aes = ggplot2::aes(),
-                   draw_panel = function(data,
-                                         panel_params,
-                                         coord,
-                                         summary.fun = head,
-                                         summary.fun.args = list()) {
-                     if (is.null(summary.fun)) {
-                       header.text <- "Input 'data' to 'draw_panel()':"
-                       z <- data
-                     } else {
-                       if (is.character(summary.fun)) {
-                         header.text <- sprintf("Summary (%s) of input 'data' to 'draw_panel()':",
-                                                summary.fun)
-                       } else {
-                         header.text <- "Summary of input 'data' to 'draw_panel()':"
-                       }
-                       z <-  do.call(summary.fun, c(quote(data), summary.fun.args))
-                     }
-                     print(header.text)
-                     print(z)
+debug_draw_function <- function(data,
+                                panel_params,
+                                coord,
+                                dbgfun.data = "head",
+                                dbgfun.data.args = list(),
+                                dbgfun.params = "summary",
+                                dbgfun.params.args = list(),
+                                dbgfun.print = "print",
+                                dbgfun.print.args = list(),
+                                draw.label = "draw_function()") {
 
-                     grid::nullGrob()
-                   },
+  if (is.character(dbgfun.print)) {
+    dbgfun.print <- match.fun(dbgfun.print)
+  } else if (!is.function(dbgfun.print)) {
+    dbgfun.print <- print
+  }
+
+  if (!is.null(dbgfun.data)) {
+    if (is.character(dbgfun.data)) {
+      dbgfun.data.name <- dbgfun.data
+      dbgfun.data <- match.fun(dbgfun.data)
+    } else if (!is.null(dbgfun.data)) {
+      dbgfun.data.name <- deparse(substitute(dbgfun.data))
+      if (length(dbgfun.data.name) > 1) {
+        dbgfun.data.name <- "anonymous function"
+      }
+    }
+    data.header <- sprintf("PANEL %i; group(s) %s; '%s' input 'data' (%s):",
+                           unclass(data$PANEL[1]),
+                           paste(format(sort(unique(data$group))), collapse = ", "),
+                           draw.label,
+                           dbgfun.data.name)
+    zz <-  do.call(dbgfun.data, c(quote(data), dbgfun.data.args))
+    dbgfun.print(data.header)
+    dbgfun.print(zz)
+  }
+
+  if (!is.null(dbgfun.params)) {
+    if (is.character(dbgfun.params)) {
+      dbgfun.params.name <- dbgfun.params
+      dbgfun.params <- match.fun(dbgfun.params)
+    } else if (!is.null(dbgfun.params)) {
+      dbgfun.params.name <- deparse(substitute(dbgfun.params))
+    }
+    params.header <- sprintf("PANEL %i; group(s) %s; '%s' input 'params' (%s):",
+                             unclass(data$PANEL[1]),
+                             paste(format(sort(unique(data$group))), collapse = ", "),
+                             draw.label,
+                             dbgfun.params.name)
+    zz <-  do.call(dbgfun.params, c(quote(panel_params), dbgfun.params.args))
+    dbgfun.print(params.header)
+    dbgfun.print(zz)
+  }
+
+  grid::nullGrob()
+}
+
+#' @rdname gginnards-ggproto
+#' @format NULL
+#' @usage NULL
+#'
+#' @export
+GeomDebugPanel <-
+  ggplot2::ggproto("GeomDebugPanel", ggplot2::Geom,
+                   # needed to avoid warnings
+                   optional_aes = optional.aes,
+                   default_aes = ggplot2::aes(),
+                   draw_panel = debug_draw_function,
                    draw_key = function(...) {
                      grid::nullGrob()
                    }
   )
+
+#' @rdname geom_debug_panel
+#'
+#' @export
+#'
+geom_debug_group <- function(mapping = NULL,
+                             data = NULL,
+                             stat = "identity",
+                             dbgfun.data = "head",
+                             dbgfun.data.args = list(),
+                             dbgfun.params = "summary",
+                             dbgfun.params.args = list(),
+                             dbgfun.print = "print",
+                             dbgfun.print.args = list(),
+                             parse = NULL,
+                             orientation = NULL,
+                             nudge_x = 0,
+                             nudge_y = 0,
+                             position = "identity",
+                             na.rm = FALSE,
+                             show.legend = FALSE,
+                             inherit.aes = TRUE,
+                             ...) {
+  if (!missing(nudge_x) || !missing(nudge_y)) {
+    if (!missing(position) && position != "identity") {
+      rlang::abort("You must specify either `position` or `nudge_x`/`nudge_y`.")
+    }
+    # by default we keep the original positions
+    position <- ggplot2::position_nudge(nudge_x, nudge_y)
+  }
+
+  ggplot2::layer(
+    geom = GeomDebugGroup,
+    mapping = mapping,
+    data = data,
+    stat = stat,
+    position = position,
+    show.legend = show.legend,
+    inherit.aes = inherit.aes,
+    params = rlang::list2(na.rm = na.rm,
+                          ...,
+                          dbgfun.data = dbgfun.data,
+                          dbgfun.data.args = dbgfun.data.args,
+                          dbgfun.params = dbgfun.params,
+                          dbgfun.params.args = dbgfun.params.args,
+                          dbgfun.print = dbgfun.print,
+                          dbgfun.print.args = dbgfun.print.args,
+                          draw.label = "draw_group()")
+  )
+}
+
+#' @rdname gginnards-ggproto
+#' @format NULL
+#' @usage NULL
+#' @export
+GeomDebugGroup <-
+  ggplot2::ggproto("GeomDebugGroup", ggplot2::Geom,
+                   # needed to avoid warnings
+                   optional_aes = optional.aes,
+                   default_aes = ggplot2::aes(),
+                   draw_group = debug_draw_function,
+                   draw_key = function(...) {
+                     grid::nullGrob()
+                   }
+  )
+